Added in Current pregnancy to prior pregnancy functionality
Added in frontend and backend logic and error checking for turning a
current pregnancy into a prior pregnancy.

+++ iTrust/src/main/edu/ncsu/csc/itrust/model/obstetrics/ObstetricsValidator.java
@@ -92,9 +92,9 @@ public class ObstetricsValidator extends POJOValidator<ObstetricsPregnancy> {
 		}
 		
 		if ( !op.getTotalWeeksPregnant().equals( "" ) ) {
-			int totalWeeks = 0;
+			double totalWeeks = 0;
 			try {
-				totalWeeks = Integer.parseInt( op.getTotalWeeksPregnant() );
+				totalWeeks = Double.parseDouble( op.getTotalWeeksPregnant() );
 			} catch ( Exception e ) {
 				errorList.addIfNotNull( "Total Weeks Pregnant must be numeric" );
 			}
@@ -118,9 +118,9 @@ public class ObstetricsValidator extends POJOValidator<ObstetricsPregnancy> {
 		}
 		
 		if ( !op.getWeightGain().equals( "" ) ) {
-			int weight = 0;
+			double weight = 0;
 			try {
-				weight = Integer.parseInt( op.getWeightGain() );
+				weight = Double.parseDouble( op.getWeightGain() );
 			} catch ( Exception e ) {
 				errorList.addIfNotNull( "Weight Gained must be numeric" );
 			}
